move models to dir, adjust & sort imports incl cli
